move logout to navbar
