+++ src/views/patient/components/Navbar.jsx
@@ -1,14 +1,17 @@
-import logo from '/images/mobilityMate-NoBg-lg.png';
+import logo from "/images/mobilityMate-NoBg-lg.png";
+import { useNavigate } from "react-router-dom";
 
 const Navbar = ({ patient }) => {
+  const navigate = useNavigate();
   return (
     <div className="navbar h-16 bg-base-100 border-b-2 px-6">
       <div className="w-1/4 flex justify-start">
         <div className="text-xl font-bold flex gap-2">
-          Welcome Back! <span className="font-light">
+          Welcome Back!{" "}
+          <span className="font-light">
             {/* {patient && patient.name ? patient.name : ''} */}
             Carl
-            </span>
+          </span>
         </div>
       </div>
       <div className="w-2/4 flex gap-2 justify-center">
@@ -29,14 +32,14 @@ const Navbar = ({ patient }) => {
               />
             </div>
           </div>
-          {/* <ul
+          <ul
             tabIndex={0}
             className="menu menu-sm dropdown-content mt-3 z-[1] p-2 shadow bg-base-100 rounded-box w-52"
           >
             <li>
-              <a>Logout</a>
+              <a onClick={() => navigate("/")}>Logout</a>
             </li>
-          </ul> */}
+          </ul>
         </div>
       </div>
     </div>
